Remove support for Rails versions lower than 8.0 and Ruby lower than 3.2
Rails 7.2 reached end of life on 2026-08-09, and it was the only reason
the gemspec still allowed Ruby 3.1, which is EOL since 2025-03-26.
