+++ .github/workflows/test.yml
@@ -10,19 +10,8 @@ jobs:
     strategy:
       matrix:
         database: [ mysql, postgresql, sqlite3 ]
-        gemfile: [ '8.1.0', '8.0.0', '7.2.0', '7.1.0', '7.0.1' ]
-        ruby: [ '4.0', '3.4', '3.3', '3.2', '3.1', '3.0' ]
-        exclude:
-          - ruby: '3.0'
-            gemfile: '7.2.0'
-          - ruby: '3.1'
-            gemfile: '8.0.0'
-          - ruby: '3.0'
-            gemfile: '8.0.0'
-          - ruby: '3.0'
-            gemfile: '8.1.0'
-          - ruby: '3.1'
-            gemfile: '8.1.0'
+        gemfile: [ '8.1.0', '8.0.0' ]
+        ruby: [ '4.0', '3.4', '3.3', '3.2' ]
       fail-fast: false
     runs-on: ubuntu-latest
     name: ${{ matrix.ruby }} ${{ matrix.database }} rails-${{ matrix.gemfile }}
@@ -61,7 +50,7 @@ jobs:
         options: --health-cmd pg_isready --health-interval 10s --health-timeout 5s --health-retries 5
         env:
           POSTGRES_PASSWORD: postgres
-  
+
       mysql:
         image: mysql:8.4
         ports: ["3306:3306"]
