+++ Appraisals
@@ -1,23 +1,20 @@
 # frozen_string_literal: true
 
 RAILS_VERSIONS = %w[
-  7.0.1
+  8.0.0
+  8.1.0
 ]
 
 RAILS_VERSIONS.each do |version|
   appraise "rails_#{version}" do
-    gem 'activemodel', version
-    gem 'activerecord', version
+    gem 'activemodel', "~> #{version}"
+    gem 'activerecord', "~> #{version}"
 
-    gem 'sqlite3', '~> 1.4', platforms: [:ruby, :rbx]
+    gem 'sqlite3', '~> 2.2', platforms: [:ruby, :rbx]
 
     if !ENV['CI'] || %w(postgres postgresql).include?(ENV['DB'])
       group :postgres, :postgresql do
-        if version =~ /^4/
-          gem 'pg', '< 1.0', platforms: [:ruby, :rbx]
-        else
-          gem 'pg', '~> 1.1', platforms: [:ruby, :rbx]
-        end
+        gem 'pg', '~> 1.1', platforms: [:ruby, :rbx]
       end
     end
