--- gemfiles/rails_7.2.0.gemfile
@@ -1,23 +1,0 @@
-# This file was generated by Appraisal
-
-source "https://rubygems.org"
-
-gem "sqlite3", "~> 1.4", platforms: [:ruby, :rbx]
-gem "activemodel", "7.2.0"
-gem "activerecord", "7.2.0"
-
-group :mysql do
-  gem "mysql2", platforms: [:ruby, :rbx]
-end
-
-group :postgres, :postgresql do
-  gem "pg", "~> 1.1", platforms: [:ruby, :rbx]
-end
-
-platforms :jruby do
-  gem "activerecord-jdbcsqlite3-adapter", "~> 1"
-  gem "activerecord-jdbcmysql-adapter", "~> 1"
-  gem "activerecord-jdbcpostgresql-adapter", "~> 1"
-end
-
-gemspec path: "../"
