+++ globalize.gemspec
@@ -15,11 +15,11 @@ Gem::Specification.new do |s|
   s.files        = Dir['{lib/**/*,[A-Z]*}']
   s.platform     = Gem::Platform::RUBY
   s.require_path = 'lib'
-  s.required_ruby_version = '>= 3.0'
+  s.required_ruby_version = '>= 3.2'
 
-  s.add_dependency 'activesupport', '>= 7.0', '< 8.2'
-  s.add_dependency 'activerecord', '>= 7.0', '< 8.2'
-  s.add_dependency 'activemodel', '>= 7.0', '< 8.2'
+  s.add_dependency 'activesupport', '>= 8.0', '< 8.2'
+  s.add_dependency 'activerecord', '>= 8.0', '< 8.2'
+  s.add_dependency 'activemodel', '>= 8.0', '< 8.2'
   s.add_dependency 'request_store', '~> 1.0'
 
   s.add_development_dependency 'appraisal'
