+++ lib/globalize/active_record/class_methods.rb
@@ -122,12 +122,8 @@ module Globalize
       def database_connection_possible?
         begin
           # Without a connection tentative, the `connected?` function can responds with a false negative
-          if Globalize.rails_7_2?
-            connection = ::ActiveRecord::Base.lease_connection
-            connection.connect!
-          else
-            ::ActiveRecord::Base.connection
-          end
+          connection = ::ActiveRecord::Base.lease_connection
+          connection.connect!
         rescue
           # Ignore connection fail because in docker build hasn't a database connection
           nil
