+++ test/data/models/serialized_attr.rb
@@ -5,21 +5,13 @@ end
 
 class ArraySerializedAttr < ActiveRecord::Base
   self.table_name = 'serialized_attrs'
-  if Globalize.rails_7_2?
-    serialize :meta, type: Array
-  else
-    serialize :meta, Array
-  end
+  serialize :meta, type: Array
   translates :meta
 end
 
 class JSONSerializedAttr < ActiveRecord::Base
   self.table_name = 'serialized_attrs'
-  if Globalize.rails_7_2?
-    serialize :meta, coder: JSON
-  else
-    serialize :meta, JSON
-  end
+  serialize :meta, coder: JSON
   translates :meta
 end
