+++ test/data/models/serialized_hash.rb
@@ -1,8 +1,4 @@
 class SerializedHash < ActiveRecord::Base
-  if Globalize.rails_7_2?
-    serialize :meta, type: Hash
-  else
-    serialize :meta, Hash
-  end
+  serialize :meta, type: Hash
   translates :meta
 end
