+++ test/globalize/attributes_test.rb
@@ -257,20 +257,9 @@ class AttributesTest < Minitest::Spec
   describe 'serializable attribute' do
     it 'keeps track of serialized attributes between classes' do
       assert_equal UnserializedAttr.globalize_serialized_attributes, {}
-
-      if Globalize.rails_7_2?
-        assert_equal SerializedAttr.globalize_serialized_attributes, {:meta=>{}}
-        assert_equal ArraySerializedAttr.globalize_serialized_attributes, {:meta=>{:type=>Array}}
-        assert_equal JSONSerializedAttr.globalize_serialized_attributes, {:meta=>{:coder=>JSON}}
-      elsif Globalize.rails_7_1?
-        assert_equal SerializedAttr.globalize_serialized_attributes, {:meta=>{}}
-        assert_equal ArraySerializedAttr.globalize_serialized_attributes, {:meta=>[Array, {}]}
-        assert_equal JSONSerializedAttr.globalize_serialized_attributes, {:meta=>[JSON, {}]}
-      else
-        assert_equal SerializedAttr.globalize_serialized_attributes[:meta].class, ActiveRecord::Coders::YAMLColumn
-        assert_equal ArraySerializedAttr.globalize_serialized_attributes[:meta].class, ActiveRecord::Coders::YAMLColumn
-        assert_equal JSONSerializedAttr.globalize_serialized_attributes[:meta], ActiveRecord::Coders::JSON
-      end
+      assert_equal SerializedAttr.globalize_serialized_attributes, {:meta=>{}}
+      assert_equal ArraySerializedAttr.globalize_serialized_attributes, {:meta=>{:type=>Array}}
+      assert_equal JSONSerializedAttr.globalize_serialized_attributes, {:meta=>{:coder=>JSON}}
     end
 
     it 'works with default marshalling, without data' do
